+++ client_test.go
@@ -39,10 +39,16 @@ func TestUserAgentHeader(t *testing.T) {
 			},
 		}),
 	)
-	_, _ = client.Auth.Login(context.Background(), qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, _ = client.V3.Encryption.Encrypt(
+		context.Background(),
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if userAgent != fmt.Sprintf("Qanapi/Go %s", internal.PackageVersion) {
 		t.Errorf("Expected User-Agent to be correct, but got: %#v", userAgent)
 	}
@@ -67,10 +73,16 @@ func TestRetryAfter(t *testing.T) {
 			},
 		}),
 	)
-	_, err := client.Auth.Login(context.Background(), qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, err := client.V3.Encryption.Encrypt(
+		context.Background(),
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if err == nil {
 		t.Error("Expected there to be a cancel error")
 	}
@@ -106,10 +118,16 @@ func TestDeleteRetryCountHeader(t *testing.T) {
 		}),
 		option.WithHeaderDel("X-Stainless-Retry-Count"),
 	)
-	_, err := client.Auth.Login(context.Background(), qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, err := client.V3.Encryption.Encrypt(
+		context.Background(),
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if err == nil {
 		t.Error("Expected there to be a cancel error")
 	}
@@ -140,10 +158,16 @@ func TestOverwriteRetryCountHeader(t *testing.T) {
 		}),
 		option.WithHeader("X-Stainless-Retry-Count", "42"),
 	)
-	_, err := client.Auth.Login(context.Background(), qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, err := client.V3.Encryption.Encrypt(
+		context.Background(),
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if err == nil {
 		t.Error("Expected there to be a cancel error")
 	}
@@ -173,10 +197,16 @@ func TestRetryAfterMs(t *testing.T) {
 			},
 		}),
 	)
-	_, err := client.Auth.Login(context.Background(), qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, err := client.V3.Encryption.Encrypt(
+		context.Background(),
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if err == nil {
 		t.Error("Expected there to be a cancel error")
 	}
@@ -200,10 +230,16 @@ func TestContextCancel(t *testing.T) {
 	)
 	cancelCtx, cancel := context.WithCancel(context.Background())
 	cancel()
-	_, err := client.Auth.Login(cancelCtx, qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, err := client.V3.Encryption.Encrypt(
+		cancelCtx,
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if err == nil {
 		t.Error("Expected there to be a cancel error")
 	}
@@ -224,10 +260,16 @@ func TestContextCancelDelay(t *testing.T) {
 	)
 	cancelCtx, cancel := context.WithTimeout(context.Background(), 2*time.Millisecond)
 	defer cancel()
-	_, err := client.Auth.Login(cancelCtx, qanapi.AuthLoginParams{
-		Email:    "valid@email.com",
-		Password: "secret1234",
-	})
+	_, err := client.V3.Encryption.Encrypt(
+		cancelCtx,
+		"{proxy}",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"password": "secret123",
+			},
+			XQanapiFields: "password",
+		},
+	)
 	if err == nil {
 		t.Error("expected there to be a cancel error")
 	}
@@ -254,10 +296,16 @@ func TestContextDeadline(t *testing.T) {
 				},
 			}),
 		)
-		_, err := client.Auth.Login(deadlineCtx, qanapi.AuthLoginParams{
-			Email:    "valid@email.com",
-			Password: "secret1234",
-		})
+		_, err := client.V3.Encryption.Encrypt(
+			deadlineCtx,
+			"{proxy}",
+			qanapi.V3EncryptionEncryptParams{
+				Data: map[string]any{
+					"password": "secret123",
+				},
+				XQanapiFields: "password",
+			},
+		)
 		if err == nil {
 			t.Error("expected there to be a deadline error")
 		}