+++ v3encryption_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/qanapi/qanapi-sdk-golang/option"
 )
 
-func TestAPIKeyRevoke(t *testing.T) {
+func TestV3EncryptionDecrypt(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -27,7 +27,16 @@ func TestAPIKeyRevoke(t *testing.T) {
 		option.WithAPIKey("My API Key"),
 		option.WithSubdomain("My-Subdomain"),
 	)
-	_, err := client.APIKeys.Revoke(context.TODO(), "apiKey")
+	_, err := client.V3.Encryption.Decrypt(
+		context.TODO(),
+		"proxy",
+		qanapi.V3EncryptionDecryptParams{
+			Data: map[string]any{
+				"foo": "bar",
+			},
+			XQanapiFields: "x-qanapi-fields",
+		},
+	)
 	if err != nil {
 		var apierr *qanapi.Error
 		if errors.As(err, &apierr) {
@@ -37,7 +46,7 @@ func TestAPIKeyRevoke(t *testing.T) {
 	}
 }
 
-func TestAPIKeyRotate(t *testing.T) {
+func TestV3EncryptionEncryptWithOptionalParams(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -51,7 +60,17 @@ func TestAPIKeyRotate(t *testing.T) {
 		option.WithAPIKey("My API Key"),
 		option.WithSubdomain("My-Subdomain"),
 	)
-	_, err := client.APIKeys.Rotate(context.TODO(), "apiKey")
+	_, err := client.V3.Encryption.Encrypt(
+		context.TODO(),
+		"proxy",
+		qanapi.V3EncryptionEncryptParams{
+			Data: map[string]any{
+				"foo": "bar",
+			},
+			XQanapiFields:      "x-qanapi-fields",
+			XQanapiDestination: qanapi.String("x-qanapi-destination"),
+		},
+	)
 	if err != nil {
 		var apierr *qanapi.Error
 		if errors.As(err, &apierr) {